pyproject.toml
[project]
name = "fair-platform"
version = "0.4.1"
description = "Fair Platform - The complete platform with backend, frontend, and CLI"
authors = [{ name = "Allan Zapata", email = "allan.zapata@up.ac.pa" }]
readme = "README.md"
requires-python = ">=3.12"
classifiers = [
    "Programming Language :: Python :: 3",
    "Operating System :: OS Independent",
    "Development Status :: 1 - Planning"
]
license = "GPL-3.0-or-later"
license-files = ["LICENSE"]
dependencies = [
    "appdirs>=1.4.4",
    "dotenv>=0.9.9",
    "fastapi>=0.116.1",
    "pydantic[email]>=2.11.7",
    "python-jose>=3.5.0",
    "python-multipart>=0.0.20",
    "sqlalchemy>=2.0.43",
    "typer>=0.17.4",
    "uvicorn>=0.35.0",
]

[project.optional-dependencies]
test = [
    "pytest>=8.0.0",
    "pytest-asyncio>=0.21.0",
    "httpx>=0.25.0",
    "pytest-mock>=3.11.0",
    "faker>=19.0.0",
    "ruff>=0.13.2",
]

[project.scripts]
fair = "fair_platform.cli.main:app"

[build-system]
requires = ["uv_build>=0.8.17,<0.9.0"]
build-backend = "uv_build"


[[tool.uv.index]]
name = "testpypi"
url = "https://test.pypi.org/simple/"
publish-url = "https://test.pypi.org/legacy/"
explicit = true